【ver0.91】 ● ズレ修正、MoveTemps.ride_onLine_vertical() を追加。 ● それに伴い、Moving にアーム制御を追加した。 ● それに伴い、Pointers に Moter* centerArm を追加。 ● AI_answer 等の修正は未完了。

+++ ai/analyze/pattern_match/PatternMatcher.cpp
@@ -13,10 +13,10 @@ void PatternMatcher::solve_D(const int* readImg) {
      * 差が最も少なかったものを採用する。
      */
     float count_tmp;
-    for (int i=0; i<8; i++) {
+    for (int8_t i=0; i<8; i++) {
         count_tmp = 0;
         // 一つずつ検証
-        for (int px=0; px<7; px++) {
+        for (int8_t px=0; px<7; px++) {
             count_tmp += abs(ansImg_D[i][px] - readImg[px]);
         }
         // 類似度（％） に変換
@@ -39,10 +39,10 @@ void PatternMatcher::solve_A(const float* readImg) {
      * 差が最も少なかったものを採用する。
      */
     float count_tmp;
-    for (int i=0; i<8; i++) {
+    for (int8_t i=0; i<8; i++) {
         count_tmp = 0;
         // 一つずつ検証
-        for (int px=0; px<7; px++) {
+        for (int8_t px=0; px<7; px++) {
             count_tmp += abs(ansImg_A[i][px] - readImg[px]);
         }
         // 類似度（％） に変換
